@@ -1,22 +1,23 @@
 {
   "name": "gearbox",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "private": true,
   "dependencies": {
     "@fontsource/lato": "^4.5.0",
+    "@tailwindcss/forms": "^0.4.0",
     "autoprefixer": "^10.4.0",
     "formik": "^2.2.6",
-    "postcss": "^8.3.11",
+    "postcss": "^8.4.4",
     "postcss-cli": "^9.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-feather": "^2.0.9",
     "react-ga": "^3.3.0",
-    "react-multi-select-component": "^4.1.13",
-    "react-router-dom": "^5.3.0",
+    "react-multi-select-component": "^4.1.14",
+    "react-router-dom": "^6.1.0",
     "react-scripts": "^4.0.3",
     "react-tooltip": "^4.2.18",
-    "tailwindcss": "^2.2.19"
+    "tailwindcss": "^3.0.0"
   },
   "scripts": {
     "watch:tailwind": "postcss src/tailwind.css -o src/index.css --watch",
@@ -46,30 +47,28 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/addons": "^6.3.12",
+    "@storybook/addon-actions": "^6.4.9",
+    "@storybook/addon-links": "^6.4.9",
+    "@storybook/addons": "^6.4.9",
     "@storybook/preset-create-react-app": "^3.2.0",
-    "@storybook/react": "^6.3.12",
-    "@tailwindcss/forms": "^0.3.4",
-    "@testing-library/jest-dom": "^5.15.0",
+    "@storybook/react": "^6.4.9",
+    "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^26.0.24",
-    "@types/node": "^14.17.33",
-    "@types/react": "^17.0.34",
+    "@types/node": "^14.18.0",
+    "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.3.2",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
-    "concurrently": "^6.3.0",
+    "concurrently": "^6.4.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "^2.4.1",
-    "typescript": "^4.4.4"
+    "prettier": "^2.5.1",
+    "typescript": "^4.5.3"
   }
 }
